Minor Adjustments
Made the foreign-key for PartnershipNight in StatsInfo nullable again.
The database previously had data in the table already that contained
null values, so it wouldn't allow me to update it to be nullable. I had
taken out the rule temporarily - here I added it (also added back the
Azure connection string - I accidentally pushed it back up with localdb
before).